Create Alarm_Clock
Alarm clock helps to alert a person or a group of people at a specified time. It is primarily used to wake people from naps and also sometimes they are used as a reminder. In this project, we will be creating an alarm clock.The person has to enter the time in hours, minutes and also choose whether it is am or pm. After clicking on the set alarm, the alarm will be set.
NOTE : Knowledge of tkinter and functions in python is needed to go ahead with the alarm clock project.
--> Project File Structure
1. Install tkinter
2. Importing libraries
3. Initializing Main Window
4. Defining function Alarm
5. Creating Labels and buttons